feat: refactor scraper validators to inherit from BaseScraperValidator and streamline reward/penalty handling
